Address some comments
