+++ rust/sdk/examples/publish.rs
@@ -1,5 +1,4 @@
-use actyx_sdk::{service::QueryResponse, tags, Ax, AxOpts, Offset};
-use futures::stream::StreamExt;
+use actyx_sdk::{tags, Ax, AxOpts};
 
 // This example demonstrates how to publish events using the higher-level API.
 #[tokio::main]
@@ -21,24 +20,6 @@ async fn main() -> anyhow::Result<()> {
             &serde_json::json!({ "temperature": 21 }),
         )?
         .await?;
-
-    let mut query_response = service.query("FROM 'sensor:temp-sensor2'").await?;
-    let offsets = loop {
-        let result = query_response.next().await.unwrap();
-        if let QueryResponse::Offsets(offsets) = result {
-            break offsets.offsets;
-        }
-    };
-
-    println!("{:?}", offsets);
-
-    let mut query_response = service
-        .query("FROM 'temperature'")
-        .with_lower_bound(offsets.clone())
-        .await?;
-    while let Some(response) = query_response.next().await {
-        println!("{:?}", response);
-    }
 
     // Print the response
     println!("{:?}", publish_response);
